non-ocp node-exporter: use different port
Prior to this commit node-exporter and the proxy used the same listen
port. On some K8s runetimes (aks) this seems to cause clashes binding
the same port twice, even though the IPs are not the same.

To fix this the node-exporter (proper) port is set to a different port
than the proxy itself.

Signed-off-by: Jacob Baungard Hansen <jacobbaungard@redhat.com>